use ints not float64s you tool

+++ topology_test.go
@@ -5,25 +5,25 @@ import (
 )
 
 func TestTopology(t *testing.T) {
-    a0 := []float64{1, 1, 3, 2, 1}
-    a1 := []float64{2, 1, 2}
-    a2 := []float64{3, 2, 1, 2, 3, 2}
-    a3 := []float64{3, 2, 1, 2, 3, 1, 3}
+    a0 := []int{1, 1, 3, 2, 1}
+    a1 := []int{2, 1, 2}
+    a2 := []int{3, 2, 1, 2, 3, 2}
+    a3 := []int{3, 2, 1, 2, 3, 1, 3}
 
-    if t0 := topology(a0); t0 != 0.0 {
-        t.Errorf("t0 expected: %f, actual: %f\n", 0.0, t0)
+    if t0 := topology(a0); t0 != 0 {
+        t.Errorf("t0 expected: %f, actual: %f\n", 0, t0)
     }
 
-    if t1 := topology(a1); t1 != 1.0 {
-        t.Errorf("t1 expected: %f, actual: %f\n", 1.0, t1)
+    if t1 := topology(a1); t1 != 1 {
+        t.Errorf("t1 expected: %f, actual: %f\n", 1, t1)
     }
 
-    if t2 := topology(a2); t2 != 4.0 {
-        t.Errorf("t2 expected: %f, actual: %f\n", 4.0, t2)
+    if t2 := topology(a2); t2 != 4 {
+        t.Errorf("t2 expected: %f, actual: %f\n", 4, t2)
     }
 
-    if t3 := topology(a3); t3 != 6.0 {
-        t.Errorf("t3 expected: %f, actual: %f\n", 6.0, t3)
+    if t3 := topology(a3); t3 != 6 {
+        t.Errorf("t3 expected: %f, actual: %f\n", 6, t3)
     }
 
     /*
